SINGA-41 Support single node single GPU training
Remove hard code of the argument for Param::GenUpdateMsg();
Now, we have to copy the gradients from gpu to cpu, and then copy the
parameter values from cpu to gpu. It is necessary for multple gpu cards
as the gradients must be aggregated on CPU (or one GPU).
But for single GPU card, we may do some special checks to avoid the
memory copy (since the parameter updates can be done in the same GPU
card as the worker works on).